Movement update. Coordinates in scale.

+++ src/ru/reeson2003/model/characters/coordinates/Coordinate.java
@@ -1,51 +1,32 @@
 package ru.reeson2003.model.characters.coordinates;
 
-import ru.reeson2003.model.characters.WorldObject;
 
 /**
- * Created by reeson on 24.12.16.
+ * Coordinate class is immutable.
  */
 public class Coordinate {
     private int x;
     private int y;
     private int z;
 
+    /**
+     * Returns instance with X, Y, Z from parameters.
+     * @param x coordinate in "X" axis.
+     * @param y coordinate in "Y" axis.
+     * @param z coordinate in "Z" axis.
+     */
     public Coordinate(int x, int y, int z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
+    /**
+     * Returns instance with X, Y, Z initialized by 0.
+     */
     public Coordinate() {
         this(0, 0, 0);
     }
-
-//    public void setX(int x) {
-////        if(x < CoordinateConstants.X_MIN)
-////            this.x = CoordinateConstants.X_MIN;
-////        else if (x > CoordinateConstants.X_MAX)
-////            this.x = CoordinateConstants.X_MAX;
-////        else
-//            this.x = x;
-//    }
-//
-//    public void setY(int y) {
-////        if(y < CoordinateConstants.Y_MIN)
-////            this.y = CoordinateConstants.Y_MIN;
-////        else if (y > CoordinateConstants.Y_MAX)
-////            this.y = CoordinateConstants.Y_MAX;
-////        else
-//            this.y = y;
-//    }
-//
-//    public void setZ(int z) {
-////        if(z < CoordinateConstants.Z_MIN)
-////            this.z = CoordinateConstants.Z_MIN;
-////        else if (z > CoordinateConstants.Z_MAX)
-////            this.z = CoordinateConstants.Z_MAX;
-////        else
-//            this.z = z;
-//    }
 
     public int getX() {
         return x;
@@ -59,15 +40,25 @@ public class Coordinate {
         return z;
     }
 
+    /**
+     * Calculates distance from this coordinate to another.
+     * @param from coordinate, to calculate distance to.
+     * @return distance from this coordinate to another.
+     */
     public int distance(Coordinate from) {
-        double dX = this.x - from.x;
-        double dY = this.y - from.y;
-        double dZ = this.z - from.z;
+        long dX = this.x - from.x;
+        long dY = this.y - from.y;
+        long dZ = this.z - from.z;
 
         int result = (int) Math.sqrt(dX * dX + dY * dY + dZ * dZ);
         return result;
     }
 
+    /**
+     * Checks if this X, Y, Z is the same as another X, Y, Z.
+     * @param c coordinate to check.
+     * @return true if this X, Y, Z is the same as another X, Y, Z.
+     */
     public boolean equals(Coordinate c) {
         return this.x == c.x && this.y == c.y && this.z == c.z;
     }
